Refactor: Cart, Order, Payment Layer (#44)
* Refector: Cart Layer 인증 방식 Store방식으로 통일

* Feat: Status Enum English화 완료

* Fix: OrderStatus 상태변경 선언방식 변경

* Refector: Order / Payments Layer 인증 방식 Store방식으로 통일

* Refector: Cart, Order, Payment Layer Exception

* Refactor: TestCode 수정

* Fix: Swagger 연동 관련 Sort 변경

* Feat: Role에 따른 Controller 접근 권한 설정

* Fix: 주문 상태 변경 권한 재설정

* Fix: Role 설정 오류 수정

* Fix: Controller Test 삭제

+++ src/main/java/com/example/demo/cart/controller/CartController.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.UUID;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.cart.dto.CartItemAddReq;
 import com.example.demo.cart.dto.CartRes;
+import com.example.demo.cart.exception.UnauthorizedException;
 import com.example.demo.cart.service.CartService;
 import com.example.demo.user.entity.UserEntity;
 
@@ -30,28 +32,44 @@ public class CartController {
 
 	private final CartService cartService;
 
+	@PreAuthorize("hasRole('CUSTOMER') or hasRole('ADMIN')")
 	@PostMapping("/add")
 	public ResponseEntity<String> addItems(@RequestBody @Valid CartItemAddReq req,
 		@AuthenticationPrincipal UserEntity user) {
+		if (user == null) {
+			throw new UnauthorizedException();
+		}
 		cartService.addItemToCart(req, user.getUserId());
 		return ResponseEntity.ok("장바구니에 추가되었습니다.");
 	}
 
+	@PreAuthorize("hasRole('CUSTOMER') or hasRole('ADMIN')")
 	@GetMapping
 	public ResponseEntity<List<CartRes>> getMyCart(@AuthenticationPrincipal UserEntity user) {
+		if (user == null) {
+			throw new UnauthorizedException();
+		}
 		List<CartRes> myCart = cartService.getMyCart(user.getUserId());
 		return ResponseEntity.ok(myCart);
 	}
 
+	@PreAuthorize("hasRole('CUSTOMER') or hasRole('ADMIN')")
 	@PatchMapping("/items/{cartItemId}")
 	public ResponseEntity<String> updateQuantity(@PathVariable UUID cartItemId, @RequestParam int quantity,
 		@AuthenticationPrincipal UserEntity user) {
+		if (user == null) {
+			throw new UnauthorizedException();
+		}
 		cartService.updateQuantity(cartItemId, quantity, user.getUserId());
 		return ResponseEntity.ok("수량이 변경되었습니다.");
 	}
 
+	@PreAuthorize("hasRole('CUSTOMER') or hasRole('ADMIN')")
 	@DeleteMapping("/items/{cartItemId}")
 	public ResponseEntity<String> deleteItem(@PathVariable UUID cartItemId, @AuthenticationPrincipal UserEntity user) {
+		if (user == null) {
+			throw new UnauthorizedException();
+		}
 		cartService.deleteItem(cartItemId, user.getUserId());
 		return ResponseEntity.ok("장바구니 항목이 삭제되었습니다.");
 	}
